Fix embedded_resize event not being triggered when the location changes but the view size stays the same.
